use usize

+++ ext/webstorage/lib.rs
@@ -19,7 +19,7 @@ pub use rusqlite;
 #[derive(Clone)]
 struct OriginStorageDir(PathBuf);
 
-const MAX_STORAGE_BYTES: u32 = 10 * 1024 * 1024;
+const MAX_STORAGE_BYTES: usize = 10 * 1024 * 1024;
 
 pub fn init(origin_storage_dir: Option<PathBuf>) -> Extension {
   Extension::builder(env!("CARGO_PKG_NAME"))
@@ -135,7 +135,7 @@ pub fn op_webstorage_key(
 }
 
 #[inline]
-fn size_check(input: u32) -> Result<(), AnyError> {
+fn size_check(input: usize) -> Result<(), AnyError> {
   if input >= MAX_STORAGE_BYTES {
     return Err(
       deno_web::DomExceptionQuotaExceededError::new(
@@ -157,13 +157,13 @@ pub fn op_webstorage_set(
 ) -> Result<(), AnyError> {
   let conn = get_webstorage(state, persistent)?;
 
-  size_check((key.len() + value.len()) as u32)?;
+  size_check(key.len() + value.len())?;
 
   let mut stmt = conn
     .prepare_cached("SELECT SUM(pgsize) FROM dbstat WHERE name = 'data'")?;
   let size: u32 = stmt.query_row(params![], |row| row.get(0))?;
 
-  size_check(size);
+  size_check(size as usize);
 
   let mut stmt = conn
     .prepare_cached("INSERT OR REPLACE INTO data (key, value) VALUES (?, ?)")?;
